move collection of kubelet logs to invariant
